+++ DiamondsAndCrystals/Engine/SolidRenderer.cpp
@@ -13,7 +13,7 @@ SolidRenderer::~SolidRenderer()
 
 void SolidRenderer::Render(SDL_Renderer * renderer, SDL_Rect * rect, float a)
 {
-	Uint8 modA = (float)m_color.a*a;
+	Uint8 modA = (Uint8)((float)m_color.a*a);
 	SDL_SetRenderDrawColor(renderer, m_color.r, m_color.g, m_color.b, modA);
 	SDL_SetRenderDrawBlendMode(renderer, SDL_BlendMode::SDL_BLENDMODE_BLEND);
 	SDL_RenderFillRect(renderer, rect);
