+++ DiamondsAndCrystals/Engine/TextRenderer.cpp
@@ -21,7 +21,7 @@ void TextRenderer::Render(SDL_Renderer* renderer, SDL_Rect* rect, float a)
 		target.y = rect->y;
 		target.w = m_width;
 		target.h = m_height;
-		Uint8 modA = 255.0*a;
+		Uint8 modA = (Uint8)(255.0*a);
 		SDL_SetTextureAlphaMod(m_texture, modA);
 		SDL_RenderCopy(m_renderer, m_texture, NULL, &target);
 	}
